GH-148 Refactors StringSpliterator into top-level class
