+++ logbook-core/src/test/java/org/zalando/logbook/ChunkingHttpLogWriterTest.java
@@ -5,7 +5,6 @@ import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Captor;
 import org.mockito.runners.MockitoJUnitRunner;
-import org.zalando.logbook.ChunkingHttpLogWriter.StringSpliterator;
 import org.zalando.logbook.DefaultLogbook.SimplePrecorrelation;
 
 import java.io.IOException;
@@ -14,7 +13,6 @@ import java.util.List;
 import static java.util.stream.Collectors.toList;
 import static org.hamcrest.Matchers.contains;
 import static org.hamcrest.Matchers.is;
-import static org.hamcrest.Matchers.nullValue;
 import static org.junit.Assert.assertThat;
 import static org.mockito.Mockito.atLeastOnce;
 import static org.mockito.Mockito.mock;
@@ -84,20 +82,4 @@ public final class ChunkingHttpLogWriterTest {
                 .map(Correlation::getResponse)
                 .collect(toList());
     }
-
-    @Test
-    public void shouldEstimateSizeWithoutTrailingPart() {
-        assertThat(new StringSpliterator("Hello", 5).estimateSize(), is(1L));
-    }
-
-    @Test
-    public void shouldEstimateSizeWithTrailingPart() {
-        assertThat(new StringSpliterator("Hello World", 5).estimateSize(), is(3L));
-    }
-
-    @Test
-    public void shouldNotSupportPartitions() {
-        assertThat(new StringSpliterator("", 0).trySplit(), is(nullValue()));
-    }
-
 }
